% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcpu-package.r
\docType{package}
\name{getcpu-package}
\alias{getcpu-package}
\title{getcpu-package}
\description{
Micro-package to determine which CPU the calling thread is running on.
}
\author{
Drew Schmidt \email{wrathematics@gmail.com}
}
\keyword{package}
